refactor: rename manager.py to job_storage.py
