+++ packages/syft-enclave/src/syft_enclaves/client.py
@@ -8,7 +8,7 @@ from syft_client.sync.peers.peer import Peer
 from syft_client.sync.peers.peer_list import PeerList
 from syft_datasets.dataset_manager import SyftDatasetManager
 from syft_job.job import JobInfo, JobsList
-from syft_job.manager import JobRef
+from syft_job.job_storage import JobRef
 from syft_job.models import JobState, JobStatus
 
 from syft_enclaves.enclave_job_info import (
